- Added numerous ASSERTs to cover areas of code where pointers and string lengths aren't fully checked. - Modified CChar::OnTick to use a proper iterator to check expiry of entries in m_lastAttackers. - Fixed a dodgy/broken way of checking for damage flags in CChar::OnTakeDamage when COMBAT_USE_RESISTANCE is enabled. Changed it to how it looks like it was intended to be. - Fixed a number of ASSERTs which downcast to WORD to validate indexes. This can cause the assertion to incorrectly succeed in some cases. - Modified CChar::Stat__SetBase to throw an exception when the stat id is invalid, rather than continuing to set the value anyway. - Fixed incorrect checking of index in CResource::r_LoadVal for "REGEN", leading to a possible memory access violation.

+++ 0.56.b/graysvr/CClientMsg.cpp
@@ -141,7 +141,7 @@ void CClient::resendBuffs()
 	}
 
 }
-void CClient::addBuff( const WORD IconId, const DWORD ClilocOne, const DWORD ClilocTwo, const short Time, LPCTSTR* pArgs, int iArgCount)
+void CClient::addBuff( const WORD IconId, const DWORD ClilocOne, const DWORD ClilocTwo, const short Time, LPCTSTR* pArgs, size_t iArgCount)
 {
 	ADDTOCALLSTACK("CClient::addBuff");
 	if ( !IsSetOF(OF_Buffs) )
